Adding loaders on db features

+++ Sources/Cloud/Commands/Database/DatabaseDelete.swift
@@ -37,10 +37,14 @@ public final class DatabaseDelete: Command {
         console.warning("WARNING: 24 hours after running this command, your data will be deleted!. Proceed with care.")
         
         let app = try console.application(for: arguments, using: cloudFactory)
-        let db_token = try self.token(for: arguments)
+        let db_token = try ServerTokens(console).token(for: arguments, repoName: app.repoName)
         
         let token = try Token.global(with: console)
         let user = try adminApi.user.get(with: token)
+       
+        guard console.confirm("Do you want to delete your database server now?") else {
+            throw "Cancelled"
+        }
         
         try CloudRedis.deleteDBServer(
             console: self.console,
@@ -48,18 +52,6 @@ public final class DatabaseDelete: Command {
             token: db_token,
             email: user.email
         )
-    }
-    
-    private func token(for arguments: [String]) -> String {
-        let token: String
-        if let chosen = arguments.option("token") {
-            token = chosen
-        } else {
-            console.error("Please define server token")
-            exit(1)
-        }
-        console.detail("token", token)
-        return token
     }
 }
